Change SpatialEntity matrices to dmat4 and Camera state to dvec3/dmat4
- SpatialEntity: modelMatrix_, worldMatrix_ are now glm::dmat4
- SpatialEntity: getOffset/setOffset use dvec3, getBasis/setBasis use dmat3
- Camera: lookFrom_, lookTo_, lookUp_ are now dvec3
- Camera: viewMatrix_, projectionMatrix_ etc. are now dmat4
- All camera derived classes updated
- CoordinateTransformer hierarchy updated to return dmat4
- TrackballObject interface updated to dvec3
- CameraProperty updated to use dvec3/dmat4/DoubleVec3RefProperty

Agent-Logs-Url: https://github.com/inviwo/inviwo/sessions/2d668631-7819-4ed9-86d8-68721b62eaa8

+++ include/inviwo/core/datastructures/camera/orthographiccamera.h
@@ -42,8 +42,8 @@ namespace inviwo {
  */
 class IVW_CORE_API OrthographicCamera final : public Camera {
 public:
-    OrthographicCamera(vec3 lookFrom = cameradefaults::lookFrom,
-                       vec3 lookTo = cameradefaults::lookTo, vec3 lookUp = cameradefaults::lookUp,
+    OrthographicCamera(dvec3 lookFrom = cameradefaults::lookFrom,
+                       dvec3 lookTo = cameradefaults::lookTo, dvec3 lookUp = cameradefaults::lookUp,
                        float nearPlane = cameradefaults::nearPlane,
                        float farPlane = cameradefaults::farPlane,
                        float aspectRatio = cameradefaults::aspectRatio,
@@ -62,14 +62,14 @@ public:
     void setWidth(float width);
     virtual void zoom(const ZoomOptions& opts) override;
 
-    virtual vec4 getClipPosFromNormalizedDeviceCoords(const vec3& ndcCoords) const override;
+    virtual dvec4 getClipPosFromNormalizedDeviceCoords(const dvec3& ndcCoords) const override;
 
     virtual void serialize(Serializer& s) const override;
     virtual void deserialize(Deserializer& d) override;
 
 protected:
     virtual bool equal(const Camera& other) const override;
-    virtual mat4 calculateProjectionMatrix() const override;
+    virtual dmat4 calculateProjectionMatrix() const override;
 
     float width_;
 };
