+++ include/inviwo/core/datastructures/camera/perspectivecamera.h
@@ -39,8 +39,8 @@ namespace inviwo {
 
 class IVW_CORE_API PerspectiveCamera final : public Camera {
 public:
-    PerspectiveCamera(vec3 lookFrom = cameradefaults::lookFrom,
-                      vec3 lookTo = cameradefaults::lookTo, vec3 lookUp = cameradefaults::lookUp,
+    PerspectiveCamera(dvec3 lookFrom = cameradefaults::lookFrom,
+                      dvec3 lookTo = cameradefaults::lookTo, dvec3 lookUp = cameradefaults::lookUp,
                       float nearPlane = cameradefaults::nearPlane,
                       float farPlane = cameradefaults::farPlane,
                       float aspectRatio = cameradefaults::aspectRatio,
@@ -65,15 +65,18 @@ public:
 
 protected:
     virtual bool equal(const Camera& other) const override;
-    virtual mat4 calculateProjectionMatrix() const override;
+    virtual dmat4 calculateProjectionMatrix() const override;
 
     float fovy_;
 };
 
 inline float PerspectiveCamera::getFovy() const { return fovy_; }
 
-inline mat4 PerspectiveCamera::calculateProjectionMatrix() const {
-    return glm::perspective(glm::radians(fovy_), aspectRatio_, nearPlaneDist_, farPlaneDist_);
+inline dmat4 PerspectiveCamera::calculateProjectionMatrix() const {
+    return glm::perspective(static_cast<double>(glm::radians(fovy_)),
+                            static_cast<double>(aspectRatio_),
+                            static_cast<double>(nearPlaneDist_),
+                            static_cast<double>(farPlaneDist_));
 }
 
 }  // namespace inviwo
